2022 Day 1 Part 2
And now we sum the top 3 elves. I wasted a *tiny* bit of time just asserting that I was summing the right 3, but I still got it on the first try.

And this was rank 52! Probably lost time due to that verification, but still! Starting out with a double leaderboard! This is a good start!