2022 Day 3 Part 2
So now it's the same as Part 1, except we're merging 3 sets instead of 2. I did waste a bit of time thinking we were subdividing the list in three insetad of taking 3 lines at a time, but thankfully I got a runtime error rather than having to wait for a bad answer submission.

I'm *really* doing well! Rank 48 on Part 2!