2022 Day 4 Part 1
Well that was a train wreck. I initially was trying to operate on the ranges numerically but I royally screwed that up. I wanted to make sure that a was always the "bigger" range so I sorted based on a's start and b's start. If b's start was less than a's, I swapped them.

The problem there is that a and b may *start* at the same number but end at different numbers. I should have sorted based on the range lengths. Super dumb oversight...

Not only that, but I wasted 3 wrong answers before switching to try a range-based approach. Then I messed that up by not getting the full ranges (forgetting to offset the end value of the range by 1 to satisfy the range constructor) triggering a 5 minute wait. Trivial to fix, but I had to wait forever to submit the right answer finally...

Absolutely abysmal rank 6444. The first wrong answer was at 2 minutes 31 seconds and I finally got it right at 13 minutes 29 seconds. That's so painful, but at least my first answer was already after the top 100 had finished. (Not that that's much consolation...)