2022 Day 4 Part 2
So now we just need to see if any part of one range is included in the other. That's a pretty simple twist, at least with the working range-based code...

This was a still atrocious (though also significantly less so) rank 4280. Given my delta and the time my first part 1 answer was given, I *might* have had a part 2 leaderboard shot had I not wrecked part 1 so badly...