2022 Day 5 Part 1
I messed this up, but only once and not too badly. I was moving crates *from* the target stack *to* the target stack, being a no-op! If I'd pulled from the source stack initially I'd have had about a 6:16 time for 82nd on the leaderboard, but alas.

Parsing the input was definitely interesting today. I'm glad for parselib!

Ultimately got 243rd. That was a costly error, but at least nowhere's near as bad as yesterday.