2022 Day 6 Part 1
Oof, I really botched that. Using the length of a set of a slice was a great idea and got me the right part of the string super quickly, but I only added 1 instead of 4 to the index! I wasted a full minute thanks to that.

Rank 234, but if I'd used the right offset it would have been 29th. Really disheartening...