2022 Day 7 Part 1
Oo, this was interesting! Parsing a UNIX-like terminal sequence is fun, especially to reconstruct a folder structure! A little bit tedious, but not too bad.

After that it's just getting folder sizes with some memoization and a loop over the folders to get the answer. (Probably should have used a comprehension but oh well.)

And I'm back! Rank 46!