2022 Refactor Day 7
Change parse_input to get_folder_sizes which goes the full way to generating the folder to size dictionary as that's all the two parts need. Furthermore, since we know the input is well formed we don't need as much validation as I was doing. We can just iterate over the lines (simpler!), change the cwd during cd commands, and accumulate sizes into the folder size dict when processing file lines. So long as we accumulate the size into cwd *and* its parents we don't even need to handle dir lines!

As far as the parts themselves, part 1 can be condensed into a single sum comprehension which looks super clean. Part 2 is marginally more involved due to needing to compute TO_FREE, but after that it again is just one min comprehension.