2022 Day 9 Part 1
Turns out I misread the movement rules but we'll get to that in part 2.

So we have a head and a tail which need to stay together with some movement rules. Moving the head is trivial, and I guessed that either the tail was touching the head or it ended up directly behind the head based on where the head just moved. Turns out this is true which is good for me!

With that (slightly mistaken) solution I managed rank 22! Woo!