2022 Day 9 Part 2
And here's where my mistaken understanding came to bite me. That movement rule works for one rope segment, but not further than that. I had to go back and reread the movement rules to see that there were 3 cases: If the tail is in the same row or column it catches up, but otherwise it *always* moves diagonally. That's how further segments keep up!

Unfortunately when I implemented that I found that I was getting the wrong answer on the example. Why? Because I had a rope that was one segment too long! I wasted so much time debugging until I looked closer and found that. I'm not sure how much time I lost, but I may well have leaderboarded part 2 had I gotten the rope length right.

In the end I dropped to rank 825 on part 2. Ouch!