2022 Refactor Day 9 even more
When trying to go to bed I realized that this problem could be solved by getting the full sequence of head positions then iteratively working backwards to get the sequence of positions for the next segment. This is even cleaner (in my opinion)! It also allows answering other questions, such as finding out how long a rope must be before the tail doesn't move.