2022 Day 10 Part 2
Oh, so this is rasterizing on a CRT, interesting. Took me longer than I'd like to have grok'd the rasterization system but that was easy to code, just record the position at each cycle instead of weighted strengths.

After that it *was* supposed to be a matter of converting to data for my lib.ocr library to parse, but I goofed that up somehow. Not sure how yet, but it was a waste of time so I ended up producing the string and parsing it with my eyeballs. I'll have to debug the lib.ocr path later...

Anyway, I did sneak in rank 72, though I definitely could have been *significantly* higher had I just made one or two fewer mistakes...