2022 Day 11 Part 1
Well that was a bit of a hot mess. I probably should have just hardcoded the input but I went ahead with parsing anyway. I did do the risky thing of delegating to eval for the operations though which at least saved time.

Anyway, then it was a matter of writing the round runner. I did that pretty quickly actually...then got a bad answer. The problem is that I messed up the parsing and was sending all items to the true destination! Worse yet, it took me 7 minutes 36 seconds of debugging to find that! Once I did finally find that, the answer was spot on, but that was a big oof!

I got a sad rank 632 for that. It looks like I would have missed the top 100 had I parse right to begin with, but only barely...