2022 Refactor Day 11
Introducing a Monkey class and extracting a solve function makes this solution code much cleaner. Even better is the operation parsing code. It's tedious, but the solution runs much faster for part 2 thanks to not doing string parsing on every item update! The rest of the parsing code is still a bit bleh, but oh well.