2022 Refactor Day 12 some more
WIth an enhancement to the graph library I can now send multiple start states (and multiple end states) into my dijkstra implementation. This allows for solving part 2 forwards with code that actually looks a bit cleaner. This isn't a major improvement, but if it helps in the future then all the better!