2022 Day 13 Part 1
Oh hey, we have a ton of json lines as input. Good thing we have the json library for parsing.

Anyway, the real problem is doing lexicographic comparisons of the parsed lists. It took a bit longer to code that up than I'd like, but it worked right the first try!

I'm back on the leaderboard! Only rank 80, but I'll take it!