2022 Rewrite Day 14
There never ended up being a need to differentiate between rock and sand tiles so the cave should just be represented by a set.

Furthermore, someone on reddit had the observation that DPF is perfect for this problem. They're absolutely right! Why bother retracing the path of the next grain of sand when you can just pick up right before the last piece settled? Solving this with recursion dramatically speeds up both parts (not that part 1 was slow) and also simplifies the code to boot!