2022 Day 16 Part 2
And here's part 2 to keep me humble! First off I expanded my breadth first iteration to having two players. I goofed up and had the first person continually branching off though which isn't the same problem. Submitted a wrong answer based on that in just under 19 minutes, oops! (This really is just letting you both open a valve and move in the same minute.)

With the right setup the state space then exploded to a crazy amount. I went through *so* many ideas trying to solve this. I tried improving the speed of operations in the breadth first iteration, I tried converting the valve keys to bit flags to optimize the tracking of opened valves but that didn't work.

Finally I tried converting this to a graph search problem where I minimized the amount of *wasted* potential. (Pressure that *would* have been released had all valves been open.) This still ran for a long time, but did give me the answer without running out of memory.

While that was running I did realize that most of the valves having 0 flow rate means that they can be treated as open for the sake of this problem. This likely prunes the state space quite a bit, but I didn't have time to try it on the breadth first expansion and while I started it on a parallel instance of the dijkstra search the first instance got the right answer first.

I dropped *all* the way to rank 105. Narrowly missed the leaderboard, but you know what? I'm still happy. A massive PB for part 1 and I solved this monster of a problem on my own!