2022 Rewrite Day 16
The fundamental approach for part 1 isn't any different, we still march through time and compute the sea of possible states for each time step. However, this is now done much more efficiently using numpy (and many associated tricks) to iterate as fast as possible and get the maximal pressure for each combination of opened valves. We don't actually care about which valve combination was chosen for part 1, but it's quick to do the transformation and it'll be needed in part 2. Aside from that, part 1 now also takes advantage of the fact that only valves with nonzero flow are useful.

Part 2 is entirely different. Some folks on the reddit pointed out a very smart approach here. If you make one pass to see what either entity could achieve by opening sets of valves you can then search every combination of valve sets. So long as the valve set pair does not overlap then the player and elephant can be treated as if they operated independently! Since we then know the best possible pressure release for each one we can sum them to see how much total pressure was released for that valve set combination.

This ends up being fastest by choosing combinations from the precomputed list, so all we need to do is check the valve sets and maximize the pressure sums over any combinations that survive.

It still takes about 0.25 seconds for part 1 and about 1 second for part 2, but that's drastically better than what I had so I'll take it.