2022 Day 17 Part 1
This took me an annoyingly long time to grok. I was thinking the jets might be the height of the chamber, or this might be tetris, or something else. Eventually I saw the note of where rocks spawn which made everything make more sense!

There probably was a faster way to encode the rocks into code too, but at least hand-coding them worked fine. (That being said, I wonder if I *needed* the center tile in the plus block...)

Anyway, after that finally hurriedly coded up the rock falling logic. Thankfully it worked fine out of the box, it probably helped that I was using sets of known coordinates to check every move. (It's such a useful technique for these sorts of problems!)

I got rank 11! I'm really on fire, that's the third day in a row I got sub 18!