2022 Day 17 Part 2
Ah, and part 2 is giving us a stupid number of rocks to fall. This was clearly a jumpahead problem, and my initial trepidation was that the way to track repeat states would require somehow keeping a record of the shape of the top of the tower which sounds super nasty!

Anyway, I first started by just tracking the current rock and position in the jet stream. So long as those match there's a chance that we're in a cycle. That didn't match the example though, so I think I next kept a list of seen heights to keep track of the height diffs. This way we can let the tower settle into a rhythm before doing the jumpahead. Unfortunately I still didn't get the right answer for the example, it was off by 1! So I then added one and submitted thinking "Why not?" Unsurprisingly, it was off.

I then tried a bunch of things, including trying to get a handle on the tower shape (just by keeping track of how far down the top item is in each column from the top of the tower) but nothing worked! The reason? For any jumpahead you need to make sure to run any "spare" cycles after the jumpahead and I was continuing from the same rock instead of the next one!

After many bad answers (and even a 5 minute timeout) I finally got the right answer 15 minutes and 47 seconds after my first bad answer.

I fell all they way to 295! Oof! Well, at least I got one leaderboard today. Though if I didn't have that off by one error I would have been on the leaderboard for part 2 as well...