2022 Refactor Day 17
Instead of duplicating the sim code it's much cleaner to have one super loop (in a deduplicated solve method) which detects cycles, jumps ahead, and then finishes out its cycles. The main thing to avoid here is trying to jumpahead multiple times, but that's easy to avoid by setting the history object to None as a sentinel that we're in the wrap up phase.

Aside from that there's some cleanup of variable names and code, but nothing too major. The fundamental algorithm is still the same.