2022 Day 18 Part 1
Okay, so that's not too bad. We have a bunch of voxels and we need to count the faces of those voxels without neighbors. In other words, the number of voxels *without* an orthogonal neighbor. Not bad to code up!

Rank 34 for that. A bit lower than I'd hoped upon seeing the problem, but I probably got my hopes up from the last few days.