2022 Day 18 Part 2
Ah, so now we have some internal air pockets to exclude from the calculation. A flood fill will work well, but it needs a boundary for the exterior to know when it is indeed an exterior! Thankfully some coordinate boundaries can work for that pretty well.

Unfortunately I goofed up in getting those boundaries by treating the minimums as the maximums for making the ranges! That dumb mistake cost me an annoying amount of time, but once fixed I got the answer easily.

I dropped all the way to rank 168, but at least today was quick. I was worried about how late I'd be up given that I need to be to church to stream in the morning!