2022 Day 19 Part 1
Oof, this is a legit hard problem to think through. I saw that a breadth first iterative approach to find the maximum geode count bogged down with a ton of states, but I figured out how to prune that down sufficiently to converge for the various blueprints. It's...not fast, and Python doesn't let me do multithreading in one process, so I brought out the multiprocessing module for an accelerated brute force solution. It's still not fast (slowed down to the slowest blueprint) but it got me the answer at least.

This atrocious hack got me rank 38, at least. That'll help me protect my overall leaderboard spot.