2022 Day 19 Part 2
Unsurprisingly part 2 assumed that you solved part 1 intelligently. I...didn't. 32 minutes makes the state space explosion that much worse!

I'm pretty sure that the intended solution is to do a sort of reverse search (have a target number of geodes and work backwards to see if that's possible to achieve) but I was just not having success coming up with a way to do that. It's probably going to be blindingly obvious once I figure it out, but that might be an exercise for tomorrow.

That being said, I did realize that I could prune states based on their maximum potential geode count. This accelerates the part 1 solution and lets me converge for a part 2 solution! I did mess up and accidentally leave the scoring code in leading to a wrong answer (6x too big) wasting some amount of time, but at least this gave the right answer.

This doubling down on my horrible hack managed rank 374. Brute force is *not* the right solution to go fast here!