2022 Optimize Day 19
I can't believe that I missed a super duper optimization. If we have enough ore robots that every turn we get enough to build *any* robot then there's no reason to build more. That alone improves the branching factor tremendously! Even taking out a couple of the hackier heuristics and operating serially this now computes the answer for part 1 in ~7-8 seconds and the answer for part 2 in ~14-15 seconds.