2022 Day 21 Part 1
Okay, so we're essentially parsing an expression and evaluating it. Well, good thing I have no reservations about using eval for speed! I probably could have coded this a bit faster but oh well.

I did give a bad answer and lose a minute due to using / instead of // which hurt! Floats aren't integers! I have a human in the loop for submitting answers, why didn't I see the fact that it was a float and abort?! Bah...

Rank 77 on the leaderboard. I'm slipping down, but at least I'm still on there. (That being said, I did lose 50-54 spots due to that oversight!)