2022 Day 21 Part 2
Oh, we're solving for a variable now! My first thought was to do a brute force search for the right answer. Now that I'm typing this up I wonder if I should have tried to do a binary search or something...oh well. Anyway, I expect that z3 (which I have no experience in) could solve this super quickly but that is a no go. I tried to give the equation to Wolfram Alpha as well but that also didn't work. So unfortunately I was left solving this myself.

This essentially is a parse tree and one side was pre-evaluated (yay) so I could just unwind the tree and modify the target until I got the right answer which is nice. That took a lot of typing though, some library to solve an expression for a variable (again, z3 probably can do it easily) would have been so nice.

Anyway, I didn't leaderboard but I at least got rank 154. Fine by me, I was not prepared for this at all!