2022 Day 22 Part 1
That's annoying, the input is set up such that my grid parser can't parse it! Oh well, I got the grid parsing done quick enough, as well as splitting the path into segments. After that it's just a straightforward problem of following the instructions on this wonky grid and finding the wraparound point when we fall off. Honestly, the trickiest bit was making sure that I got my rotations right!

I felt like I was slow but I was 21st! That's not too bad!