2022 Day 22 Part 2
Hold on, what?! You want us to take this and fold it to find the connection points on the cube for wraparounds, which can also alter the direction we're facing? Holy crap!

Instead of trying to figure *that* craziness out in code I literally cut up a postcard in the shape of my input. Then I could figure out which edges connected and code up a wraparound function for my specific input shape. (I don't know if all inputs share the same shape or not.) That was super tedious, but much less likely for me to get stuck in a hole of code that doesn't work in the slightest.

I did have a couple bugs with the wraparound function initially. One was a miscalculation which sent you into no man's land, and another was a wrap that appeared valid but was not. Thankfully I was able to detect the latter by calling wrap again on the wrapped point (facing in the opposite direction) and verify that you landed in the starting point facing the opposite direction that you started facing. If that assertion didn't match then one of the two directions was wrong! I only had one such issue thankfully, and after fixing that (with my reference paper) I got the right answer!

I got rank 12! I knew that figuring out the wrapping by hand with paper was the way to go!