2022 Day 23 Part 1
So this is our game of life problem this year. I...didn't do super great on this! I assumed that the grid was a fixed size, assumed that the search order was static, and I initially messed up my calculation of the size of the bounding rectangle because I wasn't checking for the contents of the grid cell! (I also really should have switched to using sets but didn't...)

All told this was not a great showing from me. All the way down at rank 246, ouch!