2022 Day 24 Part 1
Oh interesting, we're doing a path search on a fixed grid with moving obstacles! Multiple bits of logic to code up, but it all worked out. I did realize halfway through coding up a graph-based approach (using Dijkstra's) that doing BFS would be smarter (since calculating the blizzard positions repeatedly is a huge waste) but I was in too deep. I kept going to get it started, then on the side was just starting to refactor to a BFS when the search ended with the right answer!

Rank 19! I'm glad going with Dijkstra's didn't hurt me too much on part 1.