2022 Day 24 Part 2
Oh, that's an interesting twist. Now the path has 3 parts! There, return, then there again! All it took was adding in a phase concept (and updating my various methods for the graph search) before letting it rip. This was bound to take longer though so again I started working on a BFS version on the side but again the search finished before I got a lot of progress. I'll have to convert to BFS later.

Sidenote: I also realized that part of the overhead was reforming the blizzard position set for every state in neighbor_fn, but I could have computed that set in step_blizzards! I wonder how much time that would have saved...

Rank 50. Oof, what a drop! That's what I get for using Dijstra's instead of BFS...