2022 Rewrite Day 24
Using breadth first search is so much cleaner and faster! Not only that, but using 3 searches in a row for part 2 drastically reduces the state space since we've pruned some known suboptimal paths. This new code only takes 0.5 seconds for part 1 and 1.5 seconds for part 2, compared to who knows what for my original code!