2022 Refactor Day 24
Instead of calculating the new blizzard positions at each step it make more sense to precompute the blizzard squares at each step. However, this can be wasteful if done for the full grid. Since each axis operates independently it makes more sense to precompute the occupied squares for each axis and make use of the far smaller periods.

While this code is now a bit longer, it also runs way faster. (About 0.19 seconds for part 1 and about 0.35-0.36 seconds for part 2.)