2022 Day 25 Part 1 (There is no Part 2)
Well that was a crazy final problem! There's probably a more elegant way to convert numbers to SNAFU (maybe an offset or something) but I'm glad I got something working. I'm going to try to revisit this to convert back to SNAFU more directly without a depth first search later.

I only got rank 215 (then 180 on part 2), but given how weird this was I'm okay with that.