2022 Rewrite and generalize Day 22 Part 2
Part 2 was the sorest part of this solution. It worked but used hardcoded assumptions about the input layout that didn't even work with the example input! A generalized cubenet folding algorithm was needed to solve this more properly.

This consists of a series of steps:
1) Convert the grid into a simplified 2D unit cubenet
2) Convert these to 3D cube faces containing position and local coordinate system info
3) Find face adjacencies and edges for folding
4) For each fold edge
    a) Shift all faces such that the fold edge is at the origin
    b) Determine the fold's local coordinate system (z aligned along the fold such that x is the "before" axis and y is the "after" axis of the immediate neighbor face)
    c) Construct a fold transform which transforms from the csys to gobal, performs a unit fold, and transforms back to the csys
    d) Perform the fold on all connected faces on the fold target side

With the 2D unit cube and an associated 3D folded cube the wrapping logic can be updated, though this again requires a few more steps
1) Find the face associated with the 2D map coordinate
2) Convert the 2D vector to 3D using the face's local coordinate system
3) Find the neighboring face by moving along the 3D vector and "into" the cube
4) Determine the next 2D vector using the target face's local coordinate system
5) Determine how the faces "line up" to figure out how the 2D map coordinates are affected moving to the target square. This is a bit intricate and is fully explained in code comments

While a lot of effort to get working, this does indeed solve the problem for both the sample and real inputs. Furthermore, it correctly computes the cubenet and solution for all 28 cubenet inputs in an extended part 3 shared on Reddit: https://www.reddit.com/r/adventofcode/comments/zuso8x/2022_day_22_part_3/

Just in case cubenet folding ever comes up again, fold_cubenet has a hopefully simple enough API to extract and re-use, though tha can be done later if such a problem ever comes up.