2022 Day 11 Bug Fix
Surprisingly it did not cause a problem on my main account input, but using the mod_by logic in part one (with the repeated divisions) can cause serious problems. The division and modulus interfere with each other as the floor division can alter the remainder in modulo space! Silly oversight on my part.