2023 Day 1 Part 2
Okay, so now we have word numbers too! Simple, replace the words with numbers and run part 1 again, right? Well, if the same letter is part of two numbers we take the *first* word for the beginning number and the *second* word for the ending number. That's...non-obvious. Plus I botched other things in trying to whip this up.

I rose to rank 1270, but that's still really not good. I hope that AoC 2023 isn't *all* like this!