2023 Day 3 Part 1
Hey, it's a grid problem! lib.grid makes that easy to parse, though I did spend a tiny bit of time double-checking some of its API. Still a net win I think.

The logic of figuring out which numbers have neighboring symbols is...interesting, mostly due to needing to determine the border. Thankfully once that was done it was a simple set operation comparing with a precomputed symbol position set. I did nearly miss the fact that numbers could appear at the end of a row, but thankfully I managed to catch that before submitting my answer.

Rank 312. I didn't think I was *that* slow, but I guess I was. Holy cow!