2023 Refactor Day 3
Parts 1 and 2 duplicated a lot of code in finding the symbol positions and number borders. This all can be done in a single parse function which is significantly cleaner. Now the part 1 and part 2 specific code is nice and tidy!

(Plus I found a typo in my original code while refactoring, glad it didn't end up being a bug!)