2023 Day 4 Part 2
Ah, so now we win copies of future scorecards. This sounds daunting initially but it's just counting multiples so it's really not too bad. I did spend a tiny bit more time double-checking that we actually start with 1 of every card as that would seriously affect the result, but other than that it was pretty standard dynamic programming.

Finally! Rank 62 on the leaderboard! That said, had I not goofed part 1 I might have been around rank 30, but at least I finally got some points.