2023 Rewrite Day 5
Parts 1 and 2 can actually use the same fundamental logic (part 1 can use ranges of length 1), the only difference being how the seed line is interpreted. Furthermore, the input gives the mappings in the correct order so there is no reason to determine the application sequence either. All that's really left is the range mapping logic from part 2, but that can be (and has been) significantly cleaned up.