2023 Day 6 Part 1
Wow, I think this may be the first problem that felt like it was an appropriate difficulty for the day! It looks daunting at first, but really it's just a walk over the times and a simple multiplication to see if we win the race. Do that for all 4 races with a simple math.prod and we have our answer!

Rank 66, at least I'm not totally washed!