2023 Day 7 Part 1
Holy cow I lost so much time. Soooooooooo much time. At a high level classifying poker hands is interesting, and in fact an interesting use case of collections.Counter. However, I botched this and lost a metric butt-ton of time. It was a combination of two things, primarily:
1) I wrote a classification algorithm for a given hand, then thought (incorrectly) that I could skip it and just compare the collections.Counter.most_common() info for the two hands. However, this is not guaranteed to correctly handle two hands of different classifications. (For example, the classification of 22334 is two pair and 88245 is one pair, but the most_common() info will sort the first hand lower since the 2 8's appear greater than 2 3's!)
2) I missed an important note in the problem where it said that ties are broken based on the order of the hand, *not* descending card power! I thought I was being smart by using collections.Counter.most_common() (which *does* sort hands based on proper poker rankings, excluding straights) but the problem wanted me to be dumber. Ugh...

Rank 3533...truly atrocious. But in my defense this game of poker is super weird.