2023 Day 7 Part 2
Okay, now J is wild. Not a major curveball, honestly. Just assign it to the most powerful card that's the most common, that's always the best choice. (If it's all wild then make 5 aces.) Took a bit to whip up (and a ton of code duplication) but at least this worked first try...

Rank 2151. Absolutely abysmal...