2023 Rewrite Day 7
There are only two differences between parts 1 and 2
1) The card ranking is different (just needs a different string)
2) Jokers need to be substituted when computing the hand type

This (along with the excessively long way that the part 1 solution was written) left a ton of room for cleanup. Instead of a hand class and comparison operator, generating tuples of the hand strength and bid will have the same effect. Plus, the classification code could be drastically simplified, I kept a lot of intermediate data around for debugging (and because I wasn't sure what all I would need as I was writing the part 1 solution). The new simplified code speaks for itself, there was a lot of waste!