2023 Day 8 Part 2
Oo, and now we start at a set of nodes and travel to a set of valid target nodes simultaneously. Obviously I need a more general step counter, so part 1 got a quick refactoring. Then since the various ghosts reach the end nodes at different times (and loop!) we'll need to synchronize them all.

Initially I just wanted some info, thinking synchronizing them may be complicated. (Just because it takes N steps to reach a target node initially doesn't mean it will take N steps to reach it a second time, the target node may link to other intermediate nodes before resuming the cycle or it may resume the cycle partway through.) But to my surprise, the least common multiple of the minimum steps for each ghost *was* the answer! That must be a nice property of the inputs because that's not guaranteed by the problem text as far as I know, but I'll take it!

I'll probably whip up a more general solution though, just so it feels better.

Rank 71, there we go!