2023 Day 9 Part 2
Okay, and now we're extrapolating backwards. I mean, just reverse the list! I probably took a *bit* too long verifying in my head that just applying the same algorithm with a reversed list would work but still only a 56 second delta. (I was worried that I got something wrong when I saw how low the answer was but on thinking about it it makes sense that it would be so low!)

Rank 98. Man was today a tight race, but at least I got *a* leaderboard!