2023 Day 10 Part 1
Oo, this is an interesting problem! A grid mapping problem, but with a number of different tile types! I did goof initially by taking *all* the neighbors of S as part of the loop rather than those that physically could connect, losing a few minutes as I went to the sample inputs and ultimately found the error. But all in all not too bad.

Rank 165. Oh well, at least it wasn't too bad!