2023 Day 10 Part 2
That's a fun twist! Space filling is pretty standard but now we have a complicated border. I did miss the note that there doesn't need to be a full tile gap for tiles to be considered outside the loop since they can squeeze through. This took me a bit to notice playing with the sample inputs, as well as some thinking in how to handle it. I don't know if my approach is the best, but doubling the coordinate space and tracking edges on the loop let me turn this back into a normal space filling problem. The only special thing is counting only real tiles (those on even x,y coordinates) but beyond that it just works.

That said, my raw solution code is such a hot mess!

Rank 108, that hurts! I was so close to leaderboarding! But at least I got it without floundering forever.