2023 Day 12 Part 1
Oh hey, this reminds me of Picross! I've done this exact thing before when I wrote a Picross solver, but I took *way* too long to write this. My primary issue was that in the len(rem_splits) == 0 case I was returning *instead* of yielding, completely breaking the logic. I don't even know how long that took to find. (Still though, this is a fun example of a recursion problem!)

A painful rank 433 given that I've done this before.