2023 Day 12 Part 2
Ah, the twist is that we're needing to expand larger things as opposed to doing something with the expanded rows. Fundamentally this is pretty similar, but the generator itself needs to be filtering based on the pattern. (Furthermore, it needs to deal with counts instead of generating all possibilities!)

Couple of minor goofs, but the biggest one is that I just multiplied the pattern by 5 instead of joining with ?'s! That cost me 2:20 to figure out...

Rank 104. I *just* missed the leaderboard! Looks like I'd have had around a rank 70-75 had I expanded the pattern properly...