2023 Refactor Day 12
Parts 1 and 2 can use the same counter, and even the same solve method if it parameterizes the number of copies to make. Furthermore, lifting the match counter to be global may speed things up further by retaining the cache between pattern matches. All this nearly halves the amount of code!