2023 Optimize Day 12
There's no need to check the entire candidate pattern every iteration of the match counter loop. Instead, the next before character can be validated each iteration of the loop and *just* the damaged (and immediately following operational) sprint characters can be validated. This, along with avoiding the final recursive call when split is empty, saves about 2/3rds of the runtime.