2023 Day 13 Part 2
Oo, now we're fixing smudges in the grids. My algorithm is trash (just brute force every option until one works) and takes a couple seconds, but at least it finishes relatively quickly! That said, I did goof and submit a bad answer due to not ensuring that the found reflection was new. But on the bright side, finding *all* reflections in my part 1 code turned out to be a benefit for part 2!

Rank 64! Finally a day with two leaderboards, and they're decent!