Improve quick documentation of z3
As an experiment I tried solving 2022 Day 15 from scratch using z3. I hit an issue with needing an abs function but not knowing how to do it! That plus a few other issues make it clear that a bit better z3 doc is needed if I have a chance of using it successfully in a problem if the need arises.