2023 Day 14 Part 2
Ah, and now we're doing a cycle of shifts a ton of times! This is clearly a cycle detection and jumpahead problem, and my grid library has an as_str method making it super easy to get a key for cycle detection!

That said, I did goof initially when writing the cycle code. The rocks rolled north and west easily enough, but not south or east. I forgot that I need to iterate in reverse order for those to work, but thankfully that didn't take long to spot. (And thankfully I was validating my cycle logic with the sample input from the start!) Once that was written it was just a matter of standard cycle detection as done in the past which worked just fine.

Rank 33! Woo!