2023 Optimize Day 16
In many ways this problem is similar to a graph traversal. In fact, the grid can be simplified to just connections between points of interest, most notably splitters. The only way for light to enter a loop is through a splitter since light enters the contraption from the outside. As such, whenever tracing the light we can simply carry on until we hit the side of a splitter and actually split, or until we exit the contraption.

Furthermore, many of the paths the light takes are the same for every entry point in part 2. As such, caching the paths speeds part 2 up immensely.