2023 Refactor and optimize Day 17
Part 1 and 2 fundamentally are pretty similar, the only difference is how many minimum steps are needed before turning and the maximum steps before turning. As such a single solve function can be extracted.

In addition, the search can be done more efficiently with a heuristic. This heuristic can be based on an unconstrained cost of moving from the tile to the end as the constrained path cannot do any better (and quite likely may do worse due to needing to turn).

Finally, fix a bug where end_fn did not verify that the path ended in 4+ steps in the same direction before reaching the final node. Surprisingly this did not cause a problem when solving, despite being noted in the problem description.