2023 Optimize Day 17
The state space can be simplified a good deal. I was already thinking about returning the first neighbor node that *could* turn, but then I saw someone on reddit comment about not even tracking how many steps in the same direction have been taken. This is possible by associating *all* nodes which are allowed to turn when looking up neighbors.

In fact, this can be extended futher by not even tracking the real direction but just whether the last move was horizontal or vertical. This further reduces the state space since it doesn't matter which direction was taken previously, just which two directions are available next.

The heuristic still speeds up the search so it's worth keeping, but nowhere's near as much as it had for the previous graphs.