2023 Day 18 Part 1
Interesting. This reminds me of Day 10, but with the extra twist of needing to generate the trench/pipe from scratch. (Plus this comes with the complication of figuring out what shapes the trench lines take in the first place!)

I went with the same "count the vertical pipes" approach as day 10 which was a bit to code, but thinking about it now (way after the fact) a simple space filling algorithm would have worked just fine. Oh well, at least I had this working relatively quickly.

Rank 242. A little slow, but not too bad considering I was using a slower technique.