2023 Day 18 Part 2
I...overcomplicated this. A lot. Like, a whole lot. I thought of trying to deal with ranges and computing what is contained in the trench. That...is very complicated, especially since I was trying to generalize the vertical pipe counting method to be based on vertical pipe ranges. It's not worth going into all the details, suffice it to say that it is *extremely* complicated with a ton of edge cases and I never figured it out.

I did finally realize that I was doing this all wrong. All we have to do is compress the coordinate space (kind of the opposite of Day 11) and keep track of how big each tile is supposed to be. Then the filling algorithm is the same as part 1 and we just need to map the filled tiles back to the expanded space to count how much lava can be stored.

Ugh, rank 2350. Absolutely abysmal.